Fixes for linter
