+++ src/pages/quotes/common/components/QuoteFooter.tsx
@@ -13,9 +13,6 @@ import { InputField, Link } from '$app/components/forms';
 import { DesignSelector } from '$app/common/generic/DesignSelector';
 import { endpoint } from '$app/common/helpers';
 import { route } from '$app/common/helpers/route';
-import { useCurrentCompany } from '$app/common/hooks/useCurrentCompany';
-import { useHandleCustomFieldChange } from '$app/common/hooks/useHandleCustomFieldChange';
-import { CustomFieldsPlanAlert } from '$app/components/CustomFieldsPlanAlert';
 import { DocumentsTable } from '$app/components/DocumentsTable';
 import { MarkdownEditor } from '$app/components/forms/MarkdownEditor';
 import Toggle from '$app/components/forms/Toggle';
@@ -25,7 +22,6 @@ import { UserSelector } from '$app/components/users/UserSelector';
 import { VendorSelector } from '$app/components/vendors/VendorSelector';
 import { useAtom } from 'jotai';
 import { Upload } from '$app/pages/settings/company/documents/components';
-import { Field } from '$app/pages/settings/custom-fields/components';
 import { useTranslation } from 'react-i18next';
 import { useQueryClient } from 'react-query';
 import { useLocation, useParams } from 'react-router-dom';
@@ -43,13 +39,10 @@ export function QuoteFooter(props: Props) {
   const { id } = useParams();
   const { handleChange, errors } = props;
 
-  const company = useCurrentCompany();
   const location = useLocation();
   const queryClient = useQueryClient();
 
   const [quote] = useAtom(quoteAtom);
-
-  const handleCustomFieldChange = useHandleCustomFieldChange();
 
   const tabs = [
     t('terms'),
