+++ src/pages/vendors/edit/components/Form.tsx
@@ -11,18 +11,15 @@
 import { Card, Element } from '$app/components/cards';
 import { Button, InputField, Link } from '$app/components/forms';
 import { useCurrentCompany } from '$app/common/hooks/useCurrentCompany';
-import { useHandleCustomFieldChange } from '$app/common/hooks/useHandleCustomFieldChange';
 import { ValidationBag } from '$app/common/interfaces/validation-bag';
 import { Vendor } from '$app/common/interfaces/vendor';
 import { VendorContact } from '$app/common/interfaces/vendor-contact';
 import { Divider } from '$app/components/cards/Divider';
 import { CountrySelector } from '$app/components/CountrySelector';
 import { CustomField } from '$app/components/CustomField';
-import { CustomFieldsPlanAlert } from '$app/components/CustomFieldsPlanAlert';
 import Toggle from '$app/components/forms/Toggle';
 import { UserSelector } from '$app/components/users/UserSelector';
 import { set } from 'lodash';
-import { Field } from '$app/pages/settings/custom-fields/components';
 import { useTranslation } from 'react-i18next';
 import { TabGroup } from '$app/components/TabGroup';
 import { MarkdownEditor } from '$app/components/forms/MarkdownEditor';
@@ -39,7 +36,6 @@ export function Form(props: Props) {
   const { vendor, setVendor, errors } = props;
 
   const company = useCurrentCompany();
-  const handleCustomFieldChange = useHandleCustomFieldChange();
 
   const handleChange = (property: keyof Vendor, value: unknown) => {
     setVendor((current) => current && { ...current, [property]: value });
